FinalProject/requirements.txt
tensorflow
pytube
youtube-transcript-api
openai-whisper
textblob
pyttsx3
